Renamed builder.ts as raffle.ts
