+++ src/danmu/index.ts
@@ -3,4 +3,4 @@ export * from './tcp-controller';
 export * from './roomid-handler';
 export * from './room-collector';
 export * from './gift-history';
-export * from './builder';
+export * from './raffle';
