+++ src/danmu/raffle.ts
